Really small update to make precentage durability working (patterns)

+++ src/main/java/net/dandielo/core/items/serialize/core/Durability.java
@@ -38,7 +38,7 @@ public class Durability extends ItemAttribute {
 		{
 			if (data.endsWith("%"))
 			{
-				durabilityPercent = Integer.parseInt(data.substring(0, data.length() - 1)) / 100.0;
+				durabilityPercent = (double) Integer.parseInt(data.substring(0, data.length() - 1)) / 100.0;
 				durability = (short) (item.getMaterial().getMaxDurability() * durabilityPercent);
 			}
 			else
